Minor behaviour change to work around a bug:
If you did a "next article" that went to the next feed, you got the second article instead of the 
first one.

So, don't do a "next article" after a "next feed".

The bad side is that you need to do two "next article"s now for some reason.

Anyway I may change to a QTreeWidget ;-)