Implemented Feed.loadFull.
When set to True, it will open the post's link directly in the webview instead of the post contents.

Useful for example for dzone, where the content has a link which takes you to dzone where you need to 
click on the same link to actually see the page.

This way you save a link ;-)