Fixed (at least I hope so) the bug where suddenly old articles would appear as new.
Here's what I found.

When an article is "fresh" and "deleted" it will not actually be deleted from the DB, because
it will reappear on the next fetch.

But on update, I was setting fresh to false, then to true for the new ones... and if that second
part failed, nothing was fresh, and back to the problem.

So, I wrapped it all in a transaction, with a rollback if it fails, and that's it.